https://sonarcloud.io/project/issues?id=codewars-solutions&issues=AW_Cip88ImHFfx4_RXf9&open=AW_Cip88ImHFfx4_RXf9

+++ kata/6 kyu/bitcoin-statistics/main/Stats.java
@@ -5,7 +5,7 @@ import static java.util.stream.Stream.of;
 import java.util.DoubleSummaryStatistics;
 
 class Stats {
-  static double[][] getMinAvgMax(int discard, double[][] data) {
+  static double[][] getMinAvgMax(long discard, double[][] data) {
     var summary = new DoubleSummaryStatistics();
     var stats = stream(data).map(a -> stream(a).limit(a.length - discard).skip(discard).summaryStatistics()).peek(summary::combine);
     return concat(stats, of(summary)).map(s -> new double[] {s.getMin(), s.getAverage(), s.getMax()}).toArray(double[][]::new);
